Try storing a closure to allow capturing variables from called loc
TODO: This doesnt compile, with captured variable doesnt live long
enough. Rather the closure being stored in a Box forces it to have
potentially static lifetime (bcas it is stored in heap), even thou
we drop the thing at end of function.

TODO: Also trying to use FnMut to allow modifying captured var,
gives error, with the stored function / closure handler not
being able to be used in mutable fashion or so, need to cross check