Make SimpCmdLineManager use FnMut by storing &mut dyn FnMut
To store a closure (which inturn is kind of a trait) as part of a
data entity, one needs to store/use it has a dyn FnMut. However
even thou I assume dyn Trait is basically 2 pointers internally,
the compiler seems to say that dyn Trait's size cant be determined.

I assume bcas the data associated with the 2 pointer ie
* the actual data in the instance of the type one is working with
* the vtable associated with the type one is working with
is tightly coupled in some ways.

NOTE: working with means dyn trait actually is just a flexible proxy
to the actual underlying data entity.

So one is required to store a pointer to the dyn Trait.

And inturn given that we want FnMut, the pointer needs to be mutable.